// src/contexts/SwapFormContext/enums.ts
export enum SwapFormMarginAction {
  ADD = 'add',
  REMOVE = 'remove',
}

export enum SwapFormSubmitButtonStates {
  APPROVE_UT_PERIPHERY = 'APPROVE_UT_PERIPHERY',
  APPROVING = 'APPROVING',
  CHECKING = 'CHECKING',
  INITIALISING = 'INITIALISING',
  ROLLOVER_TRADE = 'ROLLOVER_TRADE',
  TRADE_FIXED = 'TRADE_FIXED',
  TRADE_VARIABLE = 'TRADE_VARIABLE',
  UPDATE_POSITION = 'UPDATE_POSITION',
  UPDATE = 'UPDATE',
}

export enum SwapFormSubmitButtonHintStates {
  APPROVE_NEXT_TOKEN = 'APPROVE_NEXT_TOKEN',
  APPROVE_TOKEN = 'APPROVE_TOKEN',
  APPROVING = 'APPROVING',
  CHECKING = 'CHECKING',
  ERROR_TOKEN_APPROVAL = 'ERROR_TOKEN_APPROVAL',
  FORM_INVALID = 'FORM_INVALID',
  FORM_INVALID_BALANCE = 'FORM_INVALID_BALANCE',
  FORM_INVALID_INCOMPLETE = 'FORM_INVALID_INCOMPLETE',
  FORM_INVALID_TRADE = 'FORM_INVALID_TRADE',
  INITIALISING = 'INITIALISING',
  READY_TO_TRADE_TOKENS_APPROVED = 'READY_TO_TRADE_TOKENS_APPROVED',
  READY_TO_TRADE = 'READY_TO_TRADE',
  ADD_AND_REMOVE = 'ADD_AND_REMOVE',
  ADD_AND_ADD = 'ADD_AND_ADD',
  REMOVE_AND_REMOVE = 'REMOVE_AND_REMOVE',
  REMOVE_AND_ADD = 'REMOVE_AND_ADD',
}